code climate: refactor library
